# openapi.yaml
...
openapi: 3.0.3
info:
  title: Medium Public Unofficial API
  description: >-
    
    This is an unofficial Medium API used to fetch public data such as

    users, articles, lists, etc...





    This API is heavily inspired by [Nishu Jain's Medium API](https://mediumapi.com/), in fact it's nearly the same, I just wanted to make it open source and free.





    It is done using Medium's "official" API - https://medium.com/_/api (I quote official because there is no documentation for this API and no way of knowing it exists without using Developer Tools in the web browser) - and some web scraping.





    Useful links:



    - [API's repository](https://github.com/estebanthi/medium-api)



    - [OpenAPI specification](https://github.com/estebanthi/medium-api/blob/master/openapi.yaml)
  contact:
    email: contact@esteban-thilliez.com
  version: "1.0"
servers:
  - url: https://mediumapi.esteban-thilliez.com/
...
paths:
...
  "/articles/{articleId}":
    get:
      summary: Get Article Info
      tags:
        - Articles
      description: |-
        Returns the article related information such as *Title, Subtitle,
        Tags,  Topics, Publication, etc…*
      parameters:
        - name: articleId
          in: path
          required: true
          description: It's a unique hash id assigned to every Medium article.
          schema:
            type: string
            example: 2cadf5e2561e
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      articleId:
                        type: string
                        example: 2cadf5e2561e
                      title:
                        type: string
                        example: 6 Side Business Ideas to Start Making Extra Money
                      subtitle:
                        type: string
                        example: If you want to be free and in control of your income, it’s important to
                          start side businesses now.
                      author:
                        type: string
                        example: 24babe9bccb4
                      publishedAt:
                        type: string
                        example: 2023-06-25T17:01:33.006Z
                      updatedAt:
                        type: string
                        example: null
                      url:
                        type: string
                        example: https://medium.com/@estebanthi/6-side-business-ideas-to-start-making-extra-money-2cadf5e2561e
                      publicationId:
                        type: string
                        example: null
                      tags:
                        type: array
                        items:
                          type: string
                          example: Money
                      topics:
                        type: array
                        items:
                          type: string
                          example: Money
                      wordCount:
                        type: integer
                        format: integer
                        example: 1817
                      readingTime:
                        type: number
                        example: 7.056603773584905
                      clapCount:
                        type: integer
                        example: 1
                      voters:
                        type: integer
                        example: 1
                      responsesCount:
                        type: integer
                        format: int32
                        example: 1
                      lang:
                        type: string
                        example: en
                      requiresSubscription:
                        type: boolean
                        example: true
                      isSeries:
                        type: boolean
                        example: false
                      previewImage:
                        type: string
                        example: https://miro.medium.com/1*DIWyvacsApA4926hQUiS_A.jpeg
  "/articles/{articleId}/content":
    get:
      summary: Get Article Content
      tags:
        - Articles
      description: Returns the raw article content
      parameters:
        - name: articleId
          in: path
          required: true
          description: It's a unique hash id assigned to every Medium article.
          schema:
            type: string
            example: 2cadf5e2561e
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: string
                    example: 6 Side Business Ideas to Start Making Extra Money\n\n\n\nIn a previo...
  "/articles/{articleId}/markdown":
    get:
      summary: Get Article Markdown
      tags:
        - Articles
      description: Returns the article content (markdown)
      parameters:
        - name: articleId
          in: path
          required: true
          description: It's a unique hash id assigned to every Medium article.
          schema:
            type: string
            example: 2cadf5e2561e
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: string
                    example: |-
                      # 6 Side Business Ideas to Start Making Extra Money

                      ![](https://miro...
  "/articles/{articleId}/html":
    get:
      summary: Get Article HTML
      tags:
        - Articles
      description: Returns the article content (html)
      parameters:
        - name: articleId
          in: path
          required: true
          description: It's a unique hash id assigned to every Medium article.
          schema:
            type: string
            example: 2cadf5e2561e
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: string
                    example: <h1>6 Side Business Ideas to Start Making Extra
                      Money</h1></br></br><img src='https://miro...
  "/articles/{articleId}/responses":
    get:
      summary: Get Article Responses
      tags:
        - Articles
      description: >
        Returns the article responses
      
      
        **Note:** This endpoint returns a list of ids. You can then get the responses using the `/articles/{articleId}` endpoint.
      parameters:
        - name: articleId
          in: path
          required: true
          description: It's a unique hash id assigned to every Medium article.
          schema:
            type: string
            example: 2cadf5e2561e
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                      example: 54a7237c125f
  "/articles/{articleId}/related":
    get:
      summary: Get Related Articles
      tags:
        - Articles
      description: Returns a list of 4 related articles (from the same author or publication)
      parameters:
        - name: articleId
          in: path
          required: true
          description: It's a unique hash id assigned to every Medium article.
          schema:
            type: string
            example: 2cadf5e2561e
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: string
                      example: 2a4af7ee7f1c